requirements.txt
djangorestframework
kubernetes
